+++ ros2/src/huron_ros2/include/huron_ros2/huron_node.h
@@ -39,8 +39,8 @@ class HuronNode : public rclcpp::Node {
 
   std::vector<double> joint_position_;
   std::vector<double> joint_velocity_;
-  std::vector<float> fsr_left_;
-  std::vector<float> fsr_right_;
+  std::vector<double> fsr_left_;
+  std::vector<double> fsr_right_;
 };
 
 }  // namespace ros2
